+++ src/Discriminator/Discriminators.php
@@ -2,13 +2,18 @@
 
 namespace FiiSoft\Jackdaw\Discriminator;
 
+use FiiSoft\Jackdaw\Condition\Condition;
+use FiiSoft\Jackdaw\Discriminator\Adapter\ConditionAdapter;
+use FiiSoft\Jackdaw\Discriminator\Adapter\FilterAdapter;
+use FiiSoft\Jackdaw\Discriminator\Adapter\PredicateAdapter;
 use FiiSoft\Jackdaw\Filter\Filter;
 use FiiSoft\Jackdaw\Internal\Check;
+use FiiSoft\Jackdaw\Predicate\Predicate;
 
 final class Discriminators
 {
     /**
-     * @param Discriminator|Filter|string|callable $discriminator
+     * @param Discriminator|Condition|Predicate|Filter|string|callable $discriminator
      * @return Discriminator
      */
     public static function getAdapter($discriminator): Discriminator
@@ -19,6 +24,14 @@ final class Discriminators
     
         if ($discriminator instanceof Filter) {
             return self::filter($discriminator);
+        }
+    
+        if ($discriminator instanceof Predicate) {
+            return self::predicate($discriminator);
+        }
+    
+        if ($discriminator instanceof Condition) {
+            return self::condition($discriminator);
         }
         
         if ($discriminator instanceof Discriminator) {
@@ -40,6 +53,16 @@ final class Discriminators
     public static function filter(Filter $filter, int $mode = Check::VALUE): FilterAdapter
     {
         return new FilterAdapter($filter, $mode);
+    }
+    
+    public static function predicate(Predicate $predicate): PredicateAdapter
+    {
+        return new PredicateAdapter($predicate);
+    }
+    
+    public static function condition(Condition $condition): ConditionAdapter
+    {
+        return new ConditionAdapter($condition);
     }
     
     public static function evenOdd(int $mode = Check::VALUE): EvenOdd