+++ src/Operation/Internal/Limitable.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FiiSoft\Jackdaw\Operation\Internal;
+
+interface Limitable
+{
+    public function applyLimit(int $limit): void;
+    
+    public function limit(): int;
+}