+++ src/Operation/Operation.php
@@ -8,7 +8,11 @@ interface Operation
 {
     public function handle(Signal $signal): void;
     
-    public function setNext(Operation $next): Operation;
+    public function setNext(Operation $next, bool $direct = false): Operation;
+    
+    public function setPrev(Operation $prev): void;
+    
+    public function removeFromChain(): Operation;
     
     public function streamingFinished(Signal $signal): void;