+++ src/Operation/Tail.php
@@ -9,25 +9,29 @@ final class Tail extends BaseOperation
 {
     private \SplFixedArray $buffer;
     
-    private int $numOfItems;
+    private int $length;
     private int $index = 0;
     
-    public function __construct(int $numOfItems)
+    public function __construct(int $length)
     {
-        if ($numOfItems < 0) {
-            throw new \InvalidArgumentException('Invalid param numOfItems');
+        if ($length < 0) {
+            throw new \InvalidArgumentException('Invalid param length');
         }
     
-        $this->numOfItems = $numOfItems;
-        $this->buffer = new \SplFixedArray($numOfItems);
+        $this->length = $length;
+        $this->buffer = new \SplFixedArray($length);
     }
     
     public function handle(Signal $signal): void
     {
-        if ($this->numOfItems > 0) {
-            $this->buffer[$this->index] = $signal->item->copy();
+        if ($this->length > 0) {
+            if (isset($this->buffer[$this->index])) {
+                $signal->item->copyTo($this->buffer[$this->index]);
+            } else {
+                $this->buffer[$this->index] = $signal->item->copy();
+            }
             
-            if (++$this->index === $this->numOfItems) {
+            if (++$this->index === $this->length) {
                 $this->index = 0;
             }
         }
@@ -38,7 +42,7 @@ final class Tail extends BaseOperation
         $items = [];
         
         if ($this->buffer->count() > 0) {
-            $count = \min($this->numOfItems, $this->buffer->count());
+            $count = \min($this->length, $this->buffer->count());
             
             for ($i = 0; $i < $count; ++$i) {
                 if ($this->index === $count) {
@@ -52,5 +56,15 @@ final class Tail extends BaseOperation
         }
         
         $signal->restartFrom($this->next, $items);
+    }
+    
+    public function mergeWith(Tail $other): void
+    {
+        $this->length = \min($this->length, $other->length);
+    }
+    
+    public function length(): int
+    {
+        return $this->length;
     }
 }